add 1q synthesis tests
